% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bootmse}
\alias{bootmse}
\title{Bootstrap for MSE models}
\usage{
bootmse(object, B = 500, lambda = 0, cores = 2, seed = NULL, control = list())
}
\arguments{
\item{object}{object created with \code{\link{loglinear}}.}

\item{B}{number of bootstrap samples.}

\item{lambda}{parameter for the ridge penalty \eqn{\lambda\sum\beta^2}.}

\item{cores}{the number of cores used for parallel computing.}

\item{seed}{optional seed for reproducibility.}

\item{control}{a list with the parameters:
\itemize{
\item \code{maxit} maximum number of iteration of the EM algorithm.
\item \code{tol} stop criterion for maximization of the incomplete data log likelihood.
\item \code{mtol} stop criterion for maximization of the complete data log-likelihood.
}}
}
\value{
a list with the components:
\itemize{
\item \code{par} the parameter estimates of the bootstrap sample
\item \code{fitted} the fitted population frequencies
}
}
\description{
Non-parametric bootstrap of the parameter estimates and the fitted
frequencies. 95\% confidence intervals are computed with the percentile method
and bias-corrected accelerated bootstrap (BCa) method.
}
