% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impute}
\alias{impute}
\title{Simulated data for imputation.}
\format{
Data frame with 154 rows and 7 variables.
}
\usage{
impute
}
\description{
Simulated data for imputation.
}
\keyword{datasets}
