% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcaimp}
\alias{lcaimp}
\title{Simulated data for mese, lca and imputation.}
\format{
Data frame with 156 rows and 9 variables A-c,Freq.
}
\usage{
lcaimp
}
\description{
Simulated data for mese, lca and imputation.
}
\keyword{datasets}
