% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loglinear}
\alias{loglinear}
\title{Fit log linear models for incomplete data}
\usage{
loglinear(
  formula,
  data,
  lambda = 0,
  latlevs = NULL,
  method = 1,
  stderr = c("incomplete", "complete"),
  seed = NULL,
  printPars = TRUE,
  control = list()
)
}
\arguments{
\item{formula}{formula of the model to be fitted. See 'Details'.}

\item{data}{data frame containing the manifest variables in the model. See 'Details'.}

\item{lambda}{parameter for the ridge penalty. See 'Details'.}

\item{latlevs}{vector with the number of levels of the latent
variable(s) in the formula, if any. Defaults to 2 levels.}

\item{method}{imputation method for missing values. Default option 1 uses
matrix algebra, option 2 a loop.}

\item{stderr}{computation method for the standard errors. See 'Details'.}

\item{seed}{optional seed for the starting values.}

\item{printPars}{logical for printing the log-linear parameter estimates to
the screen.}

\item{control}{list with control parameters:
\itemize{
\item \code{maxit} the maximum number of iterations of the EM algorithm. Defaults to 2000.
\item \code{tol} convergence criterion for the incomplete data log likelihood. Defaults to 1e-5.
\item \code{mtol} convergence criterion for the complete data log likelihood in the M-step.
Defaults to 1e-5
\item \code{vlat} parameter determining the variance of the random starting values for the latent
class probabilities. The value 0 yields approximately equal probabilities, other
values increase the variance.
}}
}
\value{
A list with the following components:
\itemize{
\item \code{formula} the formula of the fitted model.
\item \code{lists} names of the population registers, if present.
\item \code{lat} names of the latent variables, if present.
\item \code{dobs} contingency table of the observed data, including sampling zeros.
\item \code{dcmp} contingency table of the complete data estimates.
\item \code{pars} parameter estimates, including SE's and z- and p-values.
\item \code{probs} tables with latent class probabilities, if applicable.
\item \code{hist} iteration history.
}
}
\description{
\code{loglinear} performs a log linear analyses with optionally
missing values imputation, and/or latent class analyses and/or multiple
systems estimation.
}
\details{
The role of the variables in the formula is determined by their names.
Variables starting with an upper case letter are interpreted as lists, except
for variables named \code{X}, \code{Y} and \code{Z}, which are interpreted  as latent variables. If
there is more than one list in the formula, MSE is performed. The imputation of
missing values is done automatically.

The data can be provided in the form of a contingency table or as individual records.
In the former case sampling zeros are automatically included in the table.

The parameter \code{lambda} can be set if the algorithm breaks down due to a
non-invertible hessian. A value as small as 1e-5 is usually sufficient to
overcome this problem.

For models with latent classes and/or imputation of missing values the default
option "incomplete" of the argument \code{stderr} correctly uses the hessian of the incomplete data
log likelihood. For complex models, however, this method is painstakingly slow,
In that case it is advisable to use the option "complete", which underestimates
the standard errors but is much faster.
}
\examples{
# An MSE model with latent classes and missing value imputation for a, b, and c.
loglinear(formula = Freq ~ X * (a + b + c) + A + B + C, data = lcaimp)


}
