% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NZage2013}
\alias{NZage2013}
\title{New Zealand data with age.}
\format{
Data frame in the form aof a contingency table
with 1929 cases and 11 variables.
}
\source{
Stats New Zealand.
}
\usage{
NZage2013
}
\description{
New Zealand data with age.
}
\keyword{datasets}
