% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzy2}
\alias{nzy2}
\title{Old New Zealand data.}
\format{
Data frame in the form aof a contingency table
with 7028 cases and 10 variables.
}
\source{
Stats New Zealand.
}
\usage{
nzy2
}
\description{
Old New Zealand data.
}
\keyword{datasets}
