% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simdat}
\alias{simdat}
\title{Data simulation}
\usage{
simdat(
  lists = list(0:1, 0:1, 0:1),
  covs = list(c(1, 2, NA), c(1, 2, NA), c(1:2, NA)),
  N = 1000,
  mean = -2,
  sd = 0.5,
  seed = NULL
)
}
\arguments{
\item{lists}{list with the levels 0,1 for the registers. If a register
contains missing, add \code{NA}. \code{NULL} if the data has no registers.}

\item{covs}{list with levels of the covariates, If a covariate contains
missings add \code{NA}.\code{NULL} if the data has no covariates.}

\item{N}{population/sample size, excluding the observations with missing values.}

\item{mean, sd}{parameters of the normal distribution from which the loglinear
parameters are drawn. See 'Details'}

\item{seed}{optional seed for reproducibility.}
}
\value{
A contingency table
}
\description{
Generates a contingency table for analysis with the function
\code{loglinear}.
}
\details{
\code{mean} and \code{sd} apply to the lists
}
